requirements.txt
anyio==3.5.0
asgiref==3.5.1
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.76.0
h11==0.13.0
idna==3.3
mongoengine==0.24.1
motor==3.0.0
pydantic==1.9.0
pymongo==4.1.1
python-dotenv==0.20.0
sniffio==1.2.0
starlette==0.18.0
typing_extensions==4.2.0
uvicorn==0.17.6
